requirements.txt
langchain
openai
python-dotenv
streamlit
duckduckgo-search
numexpr